breaking: use workbox-build
