+++ package.json
@@ -16,7 +16,7 @@
     "prepublishOnly": "npm run test && npm run clean && npm run build"
   },
   "dependencies": {
-    "sw-precache": "^5.2.1"
+    "workbox-build": "^6.1.0"
   },
   "devDependencies": {
     "@babel/cli": "^7.8.0",
