+++ src/lib/index.js
@@ -1,6 +1,8 @@
-import runSWPrecache from "./run-sw-precache";
+import runWorkBoxBuild from "./run-workbox-build";
 import injectSWRegister from "./inject-sw-register";
 
-export default function() {
-  return runSWPrecache.call(this).then(() => injectSWRegister(this.public_dir));
+export default function () {
+  return runWorkBoxBuild
+    .call(this)
+    .then(() => injectSWRegister(this.public_dir));
 }
